feat(github): create octokit abstraction
With typed rest method
