+++ scripts/deno/surface-lighthouse-results.ts
@@ -1,23 +1,8 @@
-import { Octokit } from "https://cdn.skypack.dev/octokit";
-import type { RestEndpointMethodTypes } from "https://cdn.skypack.dev/@octokit/plugin-rest-endpoint-methods?dts";
+import { octokit } from "./github.ts";
 import type { EventPayloadMap } from "https://cdn.skypack.dev/@octokit/webhooks-types?dts";
 import "https://raw.githubusercontent.com/GoogleChrome/lighthouse-ci/v0.8.2/types/assert.d.ts";
 
 /* -- Setup -- */
-
-type OctokitWithRest = {
-	rest: {
-		issues: {
-			[Method in keyof RestEndpointMethodTypes["issues"]]: (
-				arg: RestEndpointMethodTypes["issues"][Method]["parameters"]
-			) => Promise<RestEndpointMethodTypes["issues"][Method]["response"]>;
-		};
-	};
-};
-
-/** Github token for Authentication */
-const token = Deno.env.get("GITHUB_TOKEN");
-if (!token) throw new Error("Missing GITHUB_TOKEN");
 
 /** Path for workflow event */
 const path = Deno.env.get("GITHUB_EVENT_PATH");
@@ -92,9 +77,6 @@ const GIHUB_PARAMS = {
 	repo: "dotcom-rendering",
 	issue_number,
 } as const;
-
-// @ts-expect-error -- Octokit’s own types are not as good as ours
-const octokit = new Octokit({ auth: token }) as OctokitWithRest;
 
 /* -- Methods -- */
